Reorganize session storage exports
