+++ .github/workflows/ci.yml
@@ -22,7 +22,7 @@ jobs:
       # TODO reinstate these tests once we re-introduce the wrappers
       # - name: Test REST resources
       #   run: yarn test:rest:resources
-      # - name: Test SessionStorage implementations
-      #   run: yarn test:session_storage
+      - name: Test SessionStorage implementations
+        run: yarn test:session_storage
       - name: Run isomorphic adapter tests
         run: yarn test:adapters
