+++ jest.config.js
@@ -2,8 +2,7 @@ const semver = require('semver');
 
 const projects = [
   './src/__tests__/jest_projects/library.jest.config.mjs',
-  // TODO Uncomment this once we're ready to test sessions
-  // './src/__tests__/jest_projects/session_storage.jest.config.mjs',
+  './src/__tests__/jest_projects/session_storage.jest.config.mjs',
   './src/__tests__/jest_projects/eslint.jest.config.mjs',
   './src/__tests__/jest_projects/adapters.mock.jest.config.mjs',
   './src/__tests__/jest_projects/adapters.node.jest.config.mjs',
