+++ src/__tests__/test-helper.ts
@@ -2,7 +2,7 @@ import * as jose from 'jose';
 
 import {shopifyApi} from '..';
 import {ConfigParams, LATEST_API_VERSION, Shopify} from '../base-types';
-import {MemorySessionStorage} from '../session/storage/memory';
+import {MemorySessionStorage} from '../session-storage/memory';
 import {JwtPayload} from '../utils/types';
 import {getHMACKey} from '../utils/get-hmac-key';
 import {mockTestRequests} from '../adapters/mock/mock_test_requests';
