+++ src/session-storage/redis.ts
@@ -1,9 +1,9 @@
 import {createClient, RedisClientOptions} from 'redis';
 
-import {SessionInterface} from '../types';
-import {SessionStorage} from '../session_storage';
-import {sessionFromEntries, sessionEntries} from '../session-utils';
-import {createSanitizeShop} from '../../utils/shop-validator';
+import {SessionInterface} from '../session/types';
+import {SessionStorage} from '../session/session_storage';
+import {sessionFromEntries, sessionEntries} from '../session/session-utils';
+import {createSanitizeShop} from '../utils/shop-validator';
 
 type RedisClient = ReturnType<typeof createClient>;
 export interface RedisSessionStorageOptions extends RedisClientOptions {
